update: extract data.py to data/source
